pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "news-agent"
version = "0.4.0"
description = "Insurance/reinsurance news monitor with priority-based email alerts."
requires-python = ">=3.11"
dependencies = [
    "feedparser>=6.0.11",
    "pyyaml>=6.0.2",
    "pydantic>=2.9",
    "pydantic-settings>=2.5",
    "structlog>=24.4",
    "python-dotenv>=1.0",
    "anthropic>=0.40",
    "apscheduler>=3.10",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "ruff>=0.7",
]
dashboard = [
    "streamlit>=1.40",
    "pandas>=2.2",
]
nikkei = [
    "browser-use>=0.12",
    "playwright>=1.50",
]

[project.scripts]
news-agent = "news_agent.__main__:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
